+++ utilities.js
@@ -1,0 +1,36 @@
+'use strict';
+
+const setHeadersAndJson = (res, json) => {
+  res.setHeader('Content-Type', 'application/json');
+  res.setHeader('Access-Control-Allow-Origin', '*');
+  res.setHeader('Access-Control-Allow-Methods', 'GET, POST, DELETE, OPTIONS');
+  res.setHeader('Access-Control-Allow-Headers', 'DNT,X-CustomHeader,Keep-Alive,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Content-Range,Range');
+  res.setHeader('Access-Control-Expose-Headers', 'DNT,X-CustomHeader,Keep-Alive,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Content-Range,Range');
+  res.json(json);
+};
+
+const { transports, createLogger, format } = require('winston');
+
+const {
+  combine, label, printf, colorize,
+} = format;
+const transport = new transports.Console({ colorize: true });
+const logFormat = printf((info) => `[${info.label}] ${info.level}: ${info.message}`);
+const logger = createLogger({
+  format: combine(
+    colorize(),
+    label({ label: 'API' }),
+    logFormat,
+  ),
+  transports: [transport],
+});
+logger.level = process.env.LOG_LEVEL || 'error';
+
+const base = 'http://xenogelion.com/Hidden/content_creator_scraper.php';
+
+module.exports = {
+  logger,
+  setHeadersAndJson,
+  baseUrl: base,
+  token: process.env.VIDEO_API_TOKEN,
+};
